[Dev Badge Giver] Update 1.2.2
Update 1.2.2:
- ➕ | Adding a little more tips to let the user know what to do
----------
Message from the creator:
This is just a minor update. And yes, the update number is kind of weird but don't mind that. I'm finally adding a readme.md to this repository and putting update files to the release section. I will update more soon, cya when I update this bad boi.
ItzKeyCode @ 2024/2/21 5:18 P.M. (UTC+8)